Set up readthedocs in shader-slang.github.io (#71)
This change prepares this repo to be used in a readthedocs project for hosting the Slang Documentation in a single place.

The setup and theming are based on the existing SlangPy readthedocs site (https://slangpy.shader-slang.org/en/latest/), but with the primary difference being that this change also adds 2 Python scripts for post-processing the html build outputs. This post processing is necessary because Sphinx/readthedocs does not assume that relative paths without file extensions should resolve to .html link, and although there is an option to set it up so that it does make that assumption, it will not work with links with anchors.
The debug prints in the script would only be visible in the readthedocs build logs.

Currently the table of contents links to the documents in this repo, as well as the docs in the Slang Standard Module Reference, with other docs like the Slang User Guide to be added in the future. For now, those links go out to the existing docs websites on GitHub Pages.