Unhide hidden tables of contents for readthedocs (#82)
The tables of contents added to the stdlib-reference repo for use on readthedocs were not hidden on GitHub pages.
Wrapping them in HTML comments hides them, but they need to be unhidden in order for RTD to be able to use the TOCs, so this change adds a function to the RTD config file to filter out the wrapper from each doc as it gets processed.