Add stable branch syncing and webhook for readthedocs build triggering (#89)
This change adds a periodic GitHub Action that pulls the main branch, checks out the latest release of Slang for the slang submodule (instead of using the master branch), and then force pushes the updated main branch and release submodule to the branch "stable", which will be used for the stable branch of the readthedocs site.

readthedocs does not recognize pushes from GitHub Actions bots, so the final step is to post a token to the site's webhook to directly trigger a new build. This change also adds that step to the submodule sync action.