Update webhook syntax and branches (#91)
The existing readthedocs webhooks do not properly rebuild the stable branch of the docs, because if the branch is not specified, RTD instead rebuilds the "latest" branch of the docs, which is tied to branch main in this repo.

This change adds the branch name to the webhook curl. If main is pushed, latest RTD will build, and if stable is pushed, stable RTD will build. If any other branch is pushed, RTD will not do anything.