Remove redundant readthedocs triggers and check for stable changes (#101)
This change removes the redundant RTD webhook triggering that did not work properly, now that pushes to the repo from the GH actions works without a manual trigger.

It also removes a check for changes from the stable branch update, since it checks if there were any changes to main, when really it should be checking for diffs between main and stable. If after the submodules were last automatically updated there have been no additional changes to main or to the submodules, the faulty check will cause the update to stable to skip. Since there will always be a diff between main and stable (slang submodule checked out from different branch) the check, even if done properly, is redundant too.