Add Command Line Reference to readthedocs (#102)
Fixes shader-slang/slang#7021

This change adds the slangc command line reference to the ReadTheDocs site, replacing the link to the .md file in the repo.

The current doc is encoded as UTF-16LE, unlike our other docs in UTF-8, and although my other change for this issue (shader-slang/slang#7048) does have a regenerated version in UTF-8, this change also adds a pre-processing step to the RTD build that re-encodes any UTF-16LE doc to UTF-8 just in case.