Add ML section to homepage and ML page accessible from topnav (#92)
This change adds a link to a dedicated ML page in the topnav, and replaces the Slang Launch Deck button on the homepage with a section about ML capabilities.
The new ML page elaborates a bit on the SlangPy and autodiff information that we have on the homepage with some extra details pulled from other pages.