Exclude slangpy docs as we redirect to dedicated docs site (#105)
This change excludes all of the slangpy docs from the RTD build, except for the index.rst, since we are currently using a redirect to point away from there towards the dedicated SlangPy RTD site for now. Excluding these docs will provide a reduction to the site's build times while a longer-term solution (as laid out in #96) is implemented.