Port semantic and section highlighting from stdlib-reference to ReadTheDocs (#104)
Fixes #79

This change ports over the section highlighting from https://shader-slang.org/stdlib-reference/global-decls/ to the readthedocs site, as well as the semantic highlighting and styles as well. It also adjusts the relative sizing of the headings to more closely resemble that site as well. RTD does have section highlighting built-in, but not when clicking on links, only when clicking on the section's pilcrow (¶).