Remove empty reflection doc, update links to go to ReadTheDocs (#134)
The page Using the Slang Compilation API has a dead link to https://shader-slang.com/slang/docs/reflection-api, which is dead because the page is not built. The page is not built because it is not in the index file, and it was deliberately omitted from there because it is just a stub containing
TODO: write documentation on reflection API here, and update link in docs/index.md

This removes the empty doc and updates the link to point at the relevant chapter of the User Guide on ReadTheDocs, and updates other links on the site to point to ReadTheDocs.
The updated links are full URLs instead of internal filepaths because we still serve versions of those pages using Jekyll in addition to the RTD versions (and are therefore still accessible to anyone with an old link), and only the RTD versions would resolve those paths.
In the future redirects could be used on those Jekyll pages to redirect to RTD in the future, making them the only version that needs to work.