Ignore all .so files in linkcheck (#178)
This changes the .so file regex patterns in linkcheck_ignore to encompass all .so paths, instead of specific .so paths.

The Slang .so libraries have been renamed, so the current regex patterns do not ignore the bad links made from references to them in the docs. Sphinx misinterprets .so filenames as website URLs with the Somalian .so country TLD, which inevitably fail the link check, so those fake links need to be ignored.